requirements.txt
telepot
pyscreenshot
opencv-python
pynput
pyperclip
pillow
requests
aiohttp>3.9.2
pyttsx3
SpeechRecognition
flask
pyngrok
python-dotenv
pyaudio
jinja2>=3.1.3
Flask-SocketIO
ollama
pandas
numpy
scikit-learn
joblib
